Replace non signal-safe signal handlers with simple ones setting a flag
We used non signal-safe functions in signal handlers (or functions called
in signal handlers) and this caused deadlocks in some occasions. (e.g.
when the signal is received while writing to syslog, the handler would
deadlock when trying to write to syslog itself, waiting on the interrupted
call to finish.)

We now set a simple variable to notify of a received signal and process
the signal in the regular loop.