Per the RFC, we must make sure MIME boundary is preceded by CRLF
NOTE: The CRLF preceding the boundary delimiter line is conceptually
attached to the boundary so that it is possible to have a part that does
not end with a CRLF (line break). Body parts that must be considered to
end with line breaks, therefore, must have two CRLFs preceding the
boundary delimiter line, the first of which is part of the preceding body
part, and the second of which is part of the encapsulation boundary.