Reorder destinations when a job is removed from runq
When a destination has multiple jobs and the first job gets removed
from the readyQ, the next job takes its place in the queue. So if the
next job priority differs from the one that got removed, it causes the
(destination) runq list to be unordered and any later sorted insert can
be fooled by it. e.g. if a job of priority 150 completes and the next
job to that destination is of priority 201, any job with priority 200
added after that will get put in front of that destination, preventing
all other destinations with jobs of priority 150 that were later in the
queue to be pushed back even lower.

We now re-sort the destination on the runq when a job is removed and
might have been on the destination readyQ.